Making Farms objects really global.
Since we have had a new instance of every service farm
in the `commons/services/__init__.py` file, we ended up wasting
resources and avoiding the class memory mechanism.

I moved the get_instance from a normal method to a class method.
Then we call it from the class without making the new instance.

I tested it with elasticsearch and graphdb and we gain to good things:

1. the connection pool of that service is always the same for every request
2. we don't inject models again at every request

Connection and models loading is done once at the server init time.